+++ UI/src/uiBindings.tsx
@@ -8,6 +8,8 @@ export class uiBindingNames
     public static SelectedDistrict                     : string = "SelectedDistrict";
     public static DistrictInfos                        : string = "DistrictInfos";
     public static DisplayOption                        : string = "DisplayOption";
+    public static ColorOption                          : string = "ColorOption";
+    public static OneColor                             : string = "OneColor";
     public static ResourceInfos                        : string = "ResourceInfos";
 }
 
@@ -16,4 +18,6 @@ export class uiEventNames
 {
     public static SelectedDistrictChanged              : string = "SelectedDistrictChanged";
     public static DisplayOptionClicked                 : string = "DisplayOptionClicked";
+    public static ColorOptionClicked                   : string = "ColorOptionClicked";
+    public static OneColorChanged                      : string = "OneColorChanged";
 }