+++ ChangeLog.md
@@ -1,1 +1,2 @@
-﻿- Better handling of text scaling changes.
+﻿- Add ability to display all buildings with one color that you choose.
+- Save and restore Display Option between games.